add_group_component {prophet}	R Documentation

Adds a component with given name that contains all of the components in group.

Description

Adds a component with given name that contains all of the components in group.

Usage

add_group_component(components, name, group)

Arguments

components
Dataframe with components.

name
Name of new group component.

group
List of components that form the group.

Value

Dataframe with components.

[Package prophet version 0.2 Index]