add_regressor {prophet}	R Documentation

Add an additional regressor to be used for fitting and predicting.

Description

The dataframe passed to 'fit' and 'predict' will have a column with the specified name to be used as a regressor. When standardize='auto', the regressor will be standardized unless it is binary. The regression coefficient is given a prior with the specified scale parameter. Decreasing the prior scale will add additional regularization. If no prior scale is provided, holidays.prior.scale will be used.

Usage

add_regressor(m, name, prior.scale = NULL, standardize = "auto")

Arguments

m
Prophet object.

name
String name of the regressor

prior.scale
Float scale for the normal prior. If not provided, holidays.prior.scale will be used.

standardize
Bool, specify whether this regressor will be standardized prior to fitting. Can be 'auto' (standardize if not binary), True, or False.

Value

The prophet model with the regressor added.

[Package prophet version 0.2 Index]