cross_validation {prophet}	R Documentation

Cross-validation for time series.

Description

Computes forecasts from historical cutoff points. Beginning from initial, makes cutoffs with a spacing of period up to (end - horizon).

Usage

cross_validation(model, horizon, units, period = NULL, initial = NULL)

Arguments

model
Fitted Prophet model.

horizon
Integer size of the horizon

units
String unit of the horizon, e.g., "days", "secs".

period
Integer amount of time between cutoff dates. Same units as horizon. If not provided, 0.5 * horizon is used.

initial
Integer size of the first training period. If not provided, 3 * horizon is used. Same units as horizon.

Details

When period is equal to the time interval of the data, this is the technique described in https://robjhyndman.com/hyndsight/tscv/ .

Value

A dataframe with the forecast, actual value, and cutoff date.

[Package prophet version 0.2 Index]