fourier_series {prophet}	R Documentation

Provides Fourier series components with the specified frequency and order.

Description

Provides Fourier series components with the specified frequency and order.

Usage

fourier_series(dates, period, series.order)

Arguments

dates
Vector of dates.

period
Number of days of the period.

series.order
Number of components.

Value

Matrix with seasonality features.

[Package prophet version 0.2 Index]