linear_growth_init {prophet}	R Documentation

Initialize linear growth.

Description

Provides a strong initialization for linear growth by calculating the growth and offset parameters that pass the function through the first and last points in the time series.

Usage

linear_growth_init(df)

Arguments

df
Data frame with columns ds (date), y_scaled (scaled time series), and t (scaled time).

Value

A vector (k, m) with the rate (k) and offset (m) of the linear growth function.

[Package prophet version 0.2 Index]