make_all_seasonality_features {prophet}	R Documentation

Dataframe with seasonality features. Includes seasonality features, holiday features, and added regressors.

Description

Dataframe with seasonality features. Includes seasonality features, holiday features, and added regressors.

Usage

make_all_seasonality_features(m, df)

Arguments

m
Prophet object.

df
Dataframe with dates for computing seasonality features and any added regressors.

Value

List with items seasonal.features: Dataframe with regressor features, prior.scales: Array of prior scales for each colum of the features dataframe.

[Package prophet version 0.2 Index]