make_future_dataframe {prophet}	R Documentation

Make dataframe with future dates for forecasting.

Description

Make dataframe with future dates for forecasting.

Usage

make_future_dataframe(m, periods, freq = "day", include_history = TRUE)

Arguments

m
Prophet model object.

periods
Int number of periods to forecast forward.

freq
'day', 'week', 'month', 'quarter', 'year', 1(1 sec), 60(1 minute) or 3600(1 hour).

include_history
Boolean to include the historical dates in the data frame for predictions.

Value

Dataframe that extends forward from the end of m$history for the requested number of periods.

[Package prophet version 0.2 Index]