prophet {prophet}	R Documentation

Prophet forecaster.

Description

Prophet forecaster.

Usage

prophet(df = NULL, growth = "linear", changepoints = NULL,
  n.changepoints = 25, yearly.seasonality = "auto",
  weekly.seasonality = "auto", daily.seasonality = "auto",
  holidays = NULL, seasonality.prior.scale = 10,
  holidays.prior.scale = 10, changepoint.prior.scale = 0.05,
  mcmc.samples = 0, interval.width = 0.8, uncertainty.samples = 1000,
  fit = TRUE, ...)

Arguments

df
(optional) Dataframe containing the history. Must have columns ds (date type) and y, the time series. If growth is logistic, then df must also have a column cap that specifies the capacity at each ds. If not provided, then the model object will be instantiated but not fit; use fit.prophet(m, df) to fit the model.

growth
String 'linear' or 'logistic' to specify a linear or logistic trend.

changepoints
Vector of dates at which to include potential changepoints. If not specified, potential changepoints are selected automatically.

n.changepoints
Number of potential changepoints to include. Not used if input 'changepoints' is supplied. If 'changepoints' is not supplied, then n.changepoints potential changepoints are selected uniformly from the first 80 percent of df$ds.

yearly.seasonality
Fit yearly seasonality. Can be 'auto', TRUE, FALSE, or a number of Fourier terms to generate.

weekly.seasonality
Fit weekly seasonality. Can be 'auto', TRUE, FALSE, or a number of Fourier terms to generate.

daily.seasonality
Fit daily seasonality. Can be 'auto', TRUE, FALSE, or a number of Fourier terms to generate.

holidays
data frame with columns holiday (character) and ds (date type)and optionally columns lower_window and upper_window which specify a range of days around the date to be included as holidays. lower_window=-2 will include 2 days prior to the date as holidays. Also optionally can have a column prior_scale specifying the prior scale for each holiday.

seasonality.prior.scale
Parameter modulating the strength of the seasonality model. Larger values allow the model to fit larger seasonal fluctuations, smaller values dampen the seasonality. Can be specified for individual seasonalities using add_seasonality.

holidays.prior.scale
Parameter modulating the strength of the holiday components model, unless overridden in the holidays input.

changepoint.prior.scale
Parameter modulating the flexibility of the automatic changepoint selection. Large values will allow many changepoints, small values will allow few changepoints.

mcmc.samples
Integer, if greater than 0, will do full Bayesian inference with the specified number of MCMC samples. If 0, will do MAP estimation.

interval.width
Numeric, width of the uncertainty intervals provided for the forecast. If mcmc.samples=0, this will be only the uncertainty in the trend using the MAP estimate of the extrapolated generative model. If mcmc.samples>0, this will be integrated over all model parameters, which will include uncertainty in seasonality.

uncertainty.samples
Number of simulated draws used to estimate uncertainty intervals.

fit
Boolean, if FALSE the model is initialized but not fit.

...
Additional arguments, passed to fit.prophet

Value

A prophet model.

Examples

## Not run: 
history <- data.frame(ds = seq(as.Date('2015-01-01'), as.Date('2016-01-01'), by = 'd'),
                      y = sin(1:366/200) + rnorm(366)/10)
m <- prophet(history)

## End(Not run)

[Package prophet version 0.2 Index]