set_auto_seasonalities {prophet}	R Documentation

Set seasonalities that were left on auto.

Description

Turns on yearly seasonality if there is >=2 years of history. Turns on weekly seasonality if there is >=2 weeks of history, and the spacing between dates in the history is <7 days. Turns on daily seasonality if there is >=2 days of history, and the spacing between dates in the history is <1 day.

Usage

set_auto_seasonalities(m)

Arguments

m
Prophet object.

Value

The prophet model with seasonalities set.

[Package prophet version 0.2 Index]