setup_dataframe {prophet}	R Documentation

Prepare dataframe for fitting or predicting.

Description

Adds a time index and scales y. Creates auxillary columns 't', 't_ix', 'y_scaled', and 'cap_scaled'. These columns are used during both fitting and predicting.

Usage

setup_dataframe(m, df, initialize_scales = FALSE)

Arguments

m
Prophet object.

df
Data frame with columns ds, y, and cap if logistic growth. Any specified additional regressors must also be present.

initialize_scales
Boolean set scaling factors in m from df.

Value

list with items 'df' and 'm'.

[Package prophet version 0.2 Index]