simulated_historical_forecasts {prophet}	R Documentation

Simulated historical forecasts.

Description

Make forecasts from k historical cutoff points, working backwards from (end - horizon) with a spacing of period between each cutoff.

Usage

simulated_historical_forecasts(model, horizon, units, k, period = NULL)

Arguments

model
Fitted Prophet model.

horizon
Integer size of the horizon

units
String unit of the horizon, e.g., "days", "secs".

k
integer number of forecast points

period
Integer amount of time between cutoff dates. Same units as horizon. If not provided, will use 0.5 * horizon.

Value

A dataframe with the forecast, actual value, and cutoff date.

[Package prophet version 0.2 Index]