validate_column_name {prophet}	R Documentation

Validates the name of a seasonality, holiday, or regressor.

Description

Validates the name of a seasonality, holiday, or regressor.

Usage

validate_column_name(m, name, check_holidays = TRUE,
  check_seasonalities = TRUE, check_regressors = TRUE)

Arguments

m
Prophet object.

name
string

check_holidays
bool check if name already used for holiday

check_seasonalities
bool check if name already used for seasonality

check_regressors
bool check if name already used for regressor

[Package prophet version 0.2 Index]